signal-generator/requirements.txt
tradingview-scraper>=0.4.0
redis>=5.0.0
aioredis>=2.0.0
websocket-client>=1.6.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
yfinance>=0.2.0
asyncio>=3.4.3
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pytz>=2023.3